refactor(hooks): flatten vendored wardley tidy engine into hooks/
Moves the vendored tidy engine out of hooks/vendor/wardley/ into the
hooks/ directory directly, alongside tidy-wardley-labels.mjs:

  vendor/wardley/tidy.mjs                 -> wardley-tidy.mjs
  vendor/wardley/wardleyLabelPlacement.mjs -> wardley-label-placement.mjs
  vendor/wardley/tidy.test.mjs            -> tests/plugin/wardley-tidy.test.mjs

These sit next to the existing *-utils.mjs helper modules; hooks.json
only registers named hooks, so the extra modules are inert support
code. PROVENANCE.md is removed — its provenance and re-sync notes are
folded into hooks/README.md and the module header comments.
