+++ .github/workflows/lint-markdown.yml
@@ -73,4 +73,4 @@ jobs:
         run: node tests/plugin/test_validate_gov_reuse_handoff.mjs
 
       - name: Wardley label-tidy hook tests
-        run: node --test tests/plugin/tidy-wardley-labels.test.mjs arckit-claude/hooks/vendor/wardley/tidy.test.mjs
+        run: node --test tests/plugin/tidy-wardley-labels.test.mjs tests/plugin/wardley-tidy.test.mjs
