+++ arckit-claude/hooks/wardley-label-placement.mjs
@@ -1,4 +1,15 @@
 /**
+ * wardley-label-placement.mjs — pure label-placement engine for wardley maps.
+ *
+ * Vendored verbatim from mermaid's `wardleyLabelPlacement.ts`
+ * (https://github.com/mermaid-js/mermaid, MIT licence), compiled to ESM and
+ * shipped via https://github.com/tractorjuice/wardley-maps-mermaid
+ * (`tools/vendor/`, commit 9abfec6adb842266bb13c12105ab8f260397084a). Renamed
+ * from `.js` to `.mjs` so it loads as ESM. Do not hand-edit — re-sync from
+ * upstream instead (see hooks/README.md).
+ */
+
+/**
  * Estimate a label's bounding box from its text and font size.
  * Uses an average glyph-width factor for sans-serif fonts. This avoids a
  * DOM measurement round-trip and keeps placement deterministic and testable.
